Add a logstable-based rudimentary explorer
It really should have a query builder functionality, using e.g. the
loki/victorialogs `GET (/loki)/api/v1/labels` and
`GET (/loki)/api/v1/label/<label_name>/values` APIs to fetch labels and
possible values. The way it's implemented now is good, but functionally
this is no different from a logstable panel on a dashboard with the query
is marginally easier to update.
However, that is specific to the datasource, so maybe this should be a
feature of the respective datasource's query builder.